requirements.txt
apache-airflow-providers-google
pyarrow
snowflake-connector-python